refactor(audio): change lyric time_ms from i64 to i32
Updates the `LyricLine` structure in both Rust and Dart to use a 32-bit integer for `time_ms` instead of a 64-bit integer. This change includes updating the generated Flutter Rust Bridge (FRB) code to support `i32` encoding and decoding via both DCO and SSE codecs.

Additionally, prevents `RustLib.init()` from being called on web platforms to avoid initialization errors.

+++ rust/src/api/audio_info.rs
@@ -22,7 +22,7 @@ pub struct AudioInfo {
 }
 
 pub struct LyricLine {
-    pub time_ms: i64,
+    pub time_ms: i32,
     pub text: String,
 }
 
@@ -96,7 +96,7 @@ impl AudioInfo {
 
                 let text = caps.get(3).unwrap().as_str().trim().to_string();
 
-                let time_ms = (min * 60_000) + (sec * 1000.0) as i64;
+                let time_ms = (min * 60_000) + (sec * 1000.0) as i32;
                 // let duration = Duration::milliseconds(total_ms);
 
                 lyrics.push(LyricLine {
